Use existing volume mount for ct config
As we already mount the pwd in to the container running chart testing, just
reuse that instead of trying to mount another directory.

When trying to mount a different directory, ran in to an issue. We can't mount
relative paths, so if you reference a relative dir (such as `.circleci`), then
it fails.

So let's prevent that and reuse what is already there. I should have though
about this before my previous commit :/